+++ mojo/helpers/redis/__init__.py
@@ -1,2 +1,6 @@
 from .client import get_connection
 from .adapter import RedisAdapter, reset_adapter, get_adapter
+
+
+def get_client():
+    return get_connection()